fikset gruppeid er 35, ikke id fra url i canvas, og at build mappen skal lages av oss, ikke autogenereres når vi compiler som jeg antok
