+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
-    "cmake.buildDirectory": "${workspaceFolder}/build/245309_Stack_Overflowers",
-    "cmake.sourceDirectory": "${workspaceFolder}/src/245309_Stack_Overflowers",
+    "cmake.buildDirectory": "${workspaceFolder}/build/35_Stack_Overflowers",
+    "cmake.sourceDirectory": "${workspaceFolder}/src/35_Stack_Overflowers",
     "C_Cpp.errorSquiggles": "disabled",
     "files.associations": {
         "cstddef": "cpp",
