--- src/245309_Stack Overflowers/include/libc/limits.h
@@ -1,4 +1,0 @@
-#pragma once
-
-#define INT_MAX 2147483647
-
