--- src/245309_Stack Overflowers/include/libc/stdbool.h
@@ -1,5 +1,0 @@
-#pragma once 
-
-#define bool unsigned char
-#define true 1
-#define false 0
