--- src/245309_Stack Overflowers/include/libc/stddef.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#define NULL ((void*)0)
