--- src/245309_Stack Overflowers/include/libc/stdio.h
@@ -1,5 +1,0 @@
-#pragma once
-
-int putchar(int ic);
-bool print(const char* data, size_t length);
-int printf(const char* __restrict__ format, ...);
