--- src/245309_Stack Overflowers/include/libc/string.h
@@ -1,4 +1,0 @@
-#pragma once
-
-
-size_t strlen(const char* str);
