--- src/245309_Stack Overflowers/limine.cfg
@@ -1,19 +1,0 @@
-# Timeout in seconds that Limine will use before automatically booting.
-TIMEOUT=5
- 
-# The entry name that will be displayed in the boot menu.
-:UiA OS (KASLR on)
-    # We use the Limine boot protocol.
-    PROTOCOL=multiboot2
- 
-    # Path to the kernel to boot. boot:/// represents the partition on which limine.cfg is located.
-    KERNEL_PATH=boot:///kernel.bin
- 
-# Same thing, but without KASLR.
-:UiA OS (KASLR off)
-    PROTOCOL=multiboot2
- 
-    # Disable KASLR (it is enabled by default for relocatable kernels)
-    KASLR=no
- 
-    KERNEL_PATH=boot:///kernel.bin
