--- src/245309_Stack Overflowers/src/kernel.cpp
@@ -1,6 +1,0 @@
-
-
-extern "C" int kernel_main(void);
-int kernel_main(){
-    return 0;
-}
