+++ src/35_Stack_Overflowers/include/libc/limits.h
@@ -1,0 +1,4 @@
+#pragma once
+
+#define INT_MAX 2147483647
+
