+++ src/35_Stack_Overflowers/include/libc/stdbool.h
@@ -1,0 +1,5 @@
+#pragma once 
+
+#define bool unsigned char
+#define true 1
+#define false 0
