+++ src/35_Stack_Overflowers/include/libc/stddef.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define NULL ((void*)0)
