+++ src/35_Stack_Overflowers/include/libc/stdint.h
@@ -1,0 +1,10 @@
+#pragma once
+
+
+typedef long unsigned int size_t;
+typedef long unsigned int uint32_t;
+typedef unsigned short uint16_t;
+typedef unsigned char uint8_t;
+typedef long int int32_t;
+typedef short int16_t;
+typedef signed char int8_t;
