+++ src/35_Stack_Overflowers/include/libc/stdio.h
@@ -1,0 +1,5 @@
+#pragma once
+
+int putchar(int ic);
+bool print(const char* data, size_t length);
+int printf(const char* __restrict__ format, ...);
