+++ src/35_Stack_Overflowers/include/libc/string.h
@@ -1,0 +1,4 @@
+#pragma once
+
+
+size_t strlen(const char* str);
