+++ src/35_Stack_Overflowers/src/kernel.cpp
@@ -1,0 +1,6 @@
+
+
+extern "C" int kernel_main(void);
+int kernel_main(){
+    return 0;
+}
